+++ pcs_test/tier1/constraint/test_config.py
@@ -336,7 +336,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint:
                 score=-1
                 Resource Set:
-                  Resources: 'G2', 'R7'
+                  Resources: 'R7', 'G2'
                   role=Started
                 Resource Set:
                   Resources: 'B2', 'R6-clone'
@@ -350,7 +350,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint:
                 kind=Optional
                 Resource Set:
-                  Resources: 'B2', 'R6-clone'
+                  Resources: 'R6-clone', 'B2'
                   require-all=0 action=stop
                 Resource Set:
                   Resources: 'G1-clone'
@@ -362,7 +362,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint:
                 ticket=ticket2
                 Resource Set:
-                  Resources: 'B2', 'G2', 'R7'
+                  Resources: 'R7', 'B2', 'G2'
                   role=Stopped
             """
         )
@@ -404,7 +404,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint:
                 score=-1
                 Resource Set:
-                  Resources: 'G2', 'R7'
+                  Resources: 'R7', 'G2'
                   role=Started
                 Resource Set:
                   Resources: 'B2', 'R6-clone'
@@ -418,7 +418,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint:
                 kind=Optional
                 Resource Set:
-                  Resources: 'B2', 'R6-clone'
+                  Resources: 'R6-clone', 'B2'
                   require-all=0 action=stop
                 Resource Set:
                   Resources: 'G1-clone'
@@ -430,7 +430,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint:
                 ticket=ticket2
                 Resource Set:
-                  Resources: 'B2', 'G2', 'R7'
+                  Resources: 'R7', 'B2', 'G2'
                   role=Stopped
             """
         )
@@ -468,7 +468,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint: colocation_set_R7G2B2
                 score=-1
                 Resource Set: colocation_set_R7G2B2_set
-                  Resources: 'G2', 'R7'
+                  Resources: 'R7', 'G2'
                   role=Started
                 Resource Set: colocation_set_R7G2B2_set-1
                   Resources: 'B2', 'R6-clone'
@@ -479,12 +479,12 @@ class ConstraintConfigText(TestCase):
               start resource 'G2' then start resource 'B2' (id: order-G2-B2-Optional)
                 kind=Optional
             Order Set Constraints:
-              Set Constraint: order_set_B2R6-cloneSe
+              Set Constraint: order_set_R6-cloneB2Se
                 kind=Optional
-                Resource Set: order_set_B2R6-cloneSe_set
-                  Resources: 'B2', 'R6-clone'
+                Resource Set: order_set_R6-cloneB2Se_set
+                  Resources: 'R6-clone', 'B2'
                   require-all=0 action=stop
-                Resource Set: order_set_B2R6-cloneSe_set-1
+                Resource Set: order_set_R6-cloneB2Se_set-1
                   Resources: 'G1-clone'
                   sequential=0 action=promote
             Ticket Constraints:
@@ -494,7 +494,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint: ticket_set_R7B2G2
                 ticket=ticket2
                 Resource Set: ticket_set_R7B2G2_set
-                  Resources: 'B2', 'G2', 'R7'
+                  Resources: 'R7', 'B2', 'G2'
                   role=Stopped
             """
         )
@@ -536,7 +536,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint: colocation_set_R7G2B2
                 score=-1
                 Resource Set: colocation_set_R7G2B2_set
-                  Resources: 'G2', 'R7'
+                  Resources: 'R7', 'G2'
                   role=Started
                 Resource Set: colocation_set_R7G2B2_set-1
                   Resources: 'B2', 'R6-clone'
@@ -547,12 +547,12 @@ class ConstraintConfigText(TestCase):
               start resource 'G2' then start resource 'B2' (id: order-G2-B2-Optional)
                 kind=Optional
             Order Set Constraints:
-              Set Constraint: order_set_B2R6-cloneSe
+              Set Constraint: order_set_R6-cloneB2Se
                 kind=Optional
-                Resource Set: order_set_B2R6-cloneSe_set
-                  Resources: 'B2', 'R6-clone'
+                Resource Set: order_set_R6-cloneB2Se_set
+                  Resources: 'R6-clone', 'B2'
                   require-all=0 action=stop
-                Resource Set: order_set_B2R6-cloneSe_set-1
+                Resource Set: order_set_R6-cloneB2Se_set-1
                   Resources: 'G1-clone'
                   sequential=0 action=promote
             Ticket Constraints:
@@ -562,7 +562,7 @@ class ConstraintConfigText(TestCase):
               Set Constraint: ticket_set_R7B2G2
                 ticket=ticket2
                 Resource Set: ticket_set_R7B2G2_set
-                  Resources: 'B2', 'G2', 'R7'
+                  Resources: 'R7', 'B2', 'G2'
                   role=Stopped
             """
         )