+++ CHANGELOG.md
@@ -14,6 +14,9 @@
 - Upgrade CIB schema version when creating a colocation constraint with
   `influence` option to be able to create such constraint ([RHEL-84511])
 - Fixed a crash in `pcs resource | stonith list` commands ([rhbz#2458608])
+- `pcs constraint config` (and its variants for each constraint type) now list
+  resources in sets in the order defined in the CIB, instead of sorting them
+  alphabetically ([rhbz#2461143])
 
 ### Deprecated
 - Commands `pcs pcsd certkey` and `pcs pcsd sync-certificates` ([RHEL-149608])
@@ -29,6 +32,7 @@
 [RHEL-84511]: https://issues.redhat.com/browse/RHEL-84511
 [RHEL-149608]: https://issues.redhat.com/browse/RHEL-149608
 [rhbz#2458608]: https://bugzilla.redhat.com/show_bug.cgi?id=2458608
+[rhbz#2461143]: https://bugzilla.redhat.com/show_bug.cgi?id=2461143
 
 
 ## [0.12.2] - 2026-01-06